Add e2e Tests
Tests which start up TeamCity in a docker container (both agent and server),
as well as an OctopusDeploy Server.

A specific, canned, config is injected to the TeamCity, and a build is commenced,
the test then ensures (via the octopus-java-api) that the OctopusServer has the
expected documents.

remove line endings