Create e2e frameworking using dockerised TeamCity server/agent (#78)
Uses TestContainers library to create a teamcity dockerised instance which can be used for e2e testing.

In order to allow the TeamCity server to operate, canned configuration (i.e. database) and project files are unpacked into TeamCity's datadirectory prior to starting the containers.

This allows a fully-populated teamcity to be started.

The teamcity rest client is no longer being actively maintained by Jetbrains, or the community, as such, the repository has been forked, and is currently published to the Octopus maven repository (and thus the poor version naming).

The rest-client is required to:

Authorise the agent with the server
Trigger builds on the server, which results in resources being sent to the Octopus Deploy server.
This contains NO tests, only framework code required for tests to be written.