Fix e2e tests
The Teamcity Docker image runs as user "tcuser" which has a user-id
of 1000.

The e2e tests mount a host directory as a volume in the container,
such that it acts as teamcity's datadir (such that projects/plugin/db
can all be mounted).

Teamcity then writes files _back_ into this directory - which may
not be deletable by the test-runner, due to mismatch in user-id).

As such, certain parts of the datadir must be deleted from _within_
the container, as it cannot be conducted from outside without
superuser privs.